requirements.txt
requests
beautifulsoup4
lxml
yt-dlp
tqdm
langdetect
pyyaml
